+++ tests/app/test_dependencies.py
@@ -22,11 +22,11 @@ from neuroagent.app.dependencies import (
     get_agent,
     get_agent_memory,
     get_bluenaas_tool,
-    get_brain_region_resolver_tool,
     get_cell_types_kg_hierarchy,
     get_chat_agent,
     get_connection_string,
     get_electrophys_feature_tool,
+    get_entities_resolver_tool,
     get_httpx_client,
     get_kg_morpho_feature_tool,
     get_kg_token,
@@ -337,9 +337,6 @@ async def test_get_agent(monkeypatch, httpx_mock, patch_required_env):
     )
 
     language_model = get_language_model(settings)
-    bluenaas_tool = get_bluenaas_tool(
-        settings=settings, token=token, httpx_client=httpx_client
-    )
     literature_tool = get_literature_tool(
         token=token, settings=settings, httpx_client=httpx_client
     )
@@ -358,7 +355,7 @@ async def test_get_agent(monkeypatch, httpx_mock, patch_required_env):
     traces_tool = get_traces_tool(
         settings=settings, token=token, httpx_client=httpx_client
     )
-    br_resolver_tool = get_brain_region_resolver_tool(
+    entities_resolver_tool = get_entities_resolver_tool(
         token=token,
         httpx_client=httpx_client,
         settings=settings,
@@ -370,9 +367,8 @@ async def test_get_agent(monkeypatch, httpx_mock, patch_required_env):
     agent = get_agent(
         valid_project,
         llm=language_model,
-        bluenaas_tool=bluenaas_tool,
         literature_tool=literature_tool,
-        br_resolver_tool=br_resolver_tool,
+        entities_resolver_tool=entities_resolver_tool,
         morpho_tool=morpho_tool,
         morphology_feature_tool=morphology_feature_tool,
         kg_morpho_feature_tool=kg_morpho_feature_tool,
@@ -416,6 +412,9 @@ async def test_get_chat_agent(
     literature_tool = get_literature_tool(
         token=token, settings=settings, httpx_client=httpx_client
     )
+    me_model_tool = get_me_model_tool(
+        settings=settings, token=token, httpx_client=httpx_client
+    )
     morpho_tool = get_morpho_tool(
         settings=settings, token=token, httpx_client=httpx_client
     )
@@ -431,7 +430,7 @@ async def test_get_chat_agent(
     traces_tool = get_traces_tool(
         settings=settings, token=token, httpx_client=httpx_client
     )
-    br_resolver_tool = get_brain_region_resolver_tool(
+    entities_resolver_tool = get_entities_resolver_tool(
         token=token,
         httpx_client=httpx_client,
         settings=settings,
@@ -444,9 +443,10 @@ async def test_get_chat_agent(
         llm=language_model,
         bluenaas_tool=bluenaas_tool,
         literature_tool=literature_tool,
-        br_resolver_tool=br_resolver_tool,
+        entities_resolver_tool=entities_resolver_tool,
         morpho_tool=morpho_tool,
         morphology_feature_tool=morphology_feature_tool,
+        me_model_tool=me_model_tool,
         kg_morpho_feature_tool=kg_morpho_feature_tool,
         electrophys_feature_tool=electrophys_feature_tool,
         traces_tool=traces_tool,